pyproject.toml
[project]
name = "rwaengine"
version = "0.1.0"
description = "Off-chain computation engine for Black-Litterman RWA fund"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.36",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "scipy>=1.12.0",
    "loguru>=0.7.2",
    "pyportfolioopt>=1.5.5",
    "scikit-learn>=1.4.0",
    "eth-account>=0.13.7",
    "xgboost>=2.0.0",
    "google-genai>=1.61.0",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/rwaengine"]
